+++ ft_memcmp.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-int			ft_memcmp(const void *s1, const void *s2, size_t n)
+int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	int		i;
+	size_t	i;
 	char	*ptr;
 	char	*str;
